LayoutTests/imported/w3c/web-platform-tests/payment-request/payment-is-showing.https-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: SecurityError: show() must be triggered by user activation.
CONSOLE MESSAGE: Unhandled Promise Rejection: SecurityError: show() must be triggered by user activation.
CONSOLE MESSAGE: Unhandled Promise Rejection: SecurityError: show() must be triggered by user activation.
CONSOLE MESSAGE: Unhandled Promise Rejection: SecurityError: show() must be triggered by user activation.
CONSOLE MESSAGE: Unhandled Promise Rejection: SecurityError: show() must be triggered by user activation.
CONSOLE MESSAGE: Unhandled Promise Rejection: SecurityError: show() must be triggered by user activation.
CONSOLE MESSAGE: Unhandled Promise Rejection: SecurityError: show() must be triggered by user activation.
CONSOLE MESSAGE: Unhandled Promise Rejection: SecurityError: show() must be triggered by user activation.
CONSOLE MESSAGE: Unhandled Promise Rejection: SecurityError: show() must be triggered by user activation.
CONSOLE MESSAGE: Unhandled Promise Rejection: SecurityError: show() must be triggered by user activation.
CONSOLE MESSAGE: Unhandled Promise Rejection: SecurityError: show() must be triggered by user activation.
CONSOLE MESSAGE: Unhandled Promise Rejection: SecurityError: show() must be triggered by user activation.


Harness Error (FAIL), message = Unhandled rejection: show() must be triggered by user activation.

FAIL The top browsing context can only show one payment sheet at a time. promise_rejects_dom: Attempting to show a second payment request must reject. function "function () { throw e }" threw object "SecurityError: show() must be triggered by user activation." that is not a DOMException AbortError: property "code" is equal to 18, expected 20
FAIL If an iframe shows a payment request, the top-level browsing context can't also show one. promise_rejects_dom: iframe is already showing a payment request. function "function () { throw e }" threw object "SecurityError: show() must be triggered by user activation." that is not a DOMException AbortError: property "code" is equal to 18, expected 20
FAIL An iframe cannot show a payment request if the top-level window is already showing one. promise_rejects_dom: The top window is already showing a payment request. function "function () { throw e }" threw object "SecurityError: show() must be triggered by user activation." that is not a DOMException AbortError: property "code" is equal to 18, expected 20
FAIL Using a popup window prevents the top-browsing context from showing a payment request promise_rejects_dom: Expected window's showPromise to reject, request is already showing function "function () { throw e }" threw object "SecurityError: show() must be triggered by user activation." that is not a DOMException AbortError: property "code" is equal to 18, expected 20
FAIL Given multiple nested browsing contexts, and window calls show() first, other nested browsing contexts can't show a request. promise_rejects_dom: Expected popupShowPromise to reject, request is already showing. function "function () { throw e }" threw object "SecurityError: show() must be triggered by user activation." that is not a DOMException AbortError: property "code" is equal to 18, expected 20
FAIL Given multiple nested browsing contexts, and popup calls show() first, other nested browsing contexts can't show a request. promise_rejects_dom: Expected windowShowPromise to reject, the popup is showing a payment request. function "function () { throw e }" threw object "SecurityError: show() must be triggered by user activation." that is not a DOMException AbortError: property "code" is equal to 18, expected 20
FAIL Given multiple nested browsing contexts, and an iframe calls show() first, other nested browsing contexts can't show a request. promise_rejects_dom: Expected windowShowPromise to reject, the popup is showing a payment request. function "function () { throw e }" threw object "SecurityError: show() must be triggered by user activation." that is not a DOMException AbortError: property "code" is equal to 18, expected 20
FAIL Navigating an iframe as a nested browsing context sets 'payment request is showing boolean' to false. promise_rejects_dom: Navigating iframe away must cause the iframeShowPromise to reject. function "function () { throw e }" threw object "SecurityError: show() must be triggered by user activation." that is not a DOMException AbortError: property "code" is equal to 18, expected 20
TIMEOUT Navigating a popup as a nested browsing context sets 'payment request is showing boolean' to false. Test timed out

Harness Error (FAIL), message = Unhandled rejection: show() must be triggered by user activation.

FAIL The top browsing context can only show one payment sheet at a time. promise_rejects_dom: Attempting to show a second payment request must reject. function "function () { throw e }" threw object "SecurityError: show() must be triggered by user activation." that is not a DOMException AbortError: property "code" is equal to 18, expected 20
FAIL If an iframe shows a payment request, the top-level browsing context can't also show one. promise_rejects_dom: iframe is already showing a payment request. function "function () { throw e }" threw object "SecurityError: show() must be triggered by user activation." that is not a DOMException AbortError: property "code" is equal to 18, expected 20
FAIL An iframe cannot show a payment request if the top-level window is already showing one. promise_rejects_dom: The top window is already showing a payment request. function "function () { throw e }" threw object "SecurityError: show() must be triggered by user activation." that is not a DOMException AbortError: property "code" is equal to 18, expected 20
FAIL Using a popup window prevents the top-browsing context from showing a payment request promise_rejects_dom: Expected window's showPromise to reject, request is already showing function "function () { throw e }" threw object "SecurityError: show() must be triggered by user activation." that is not a DOMException AbortError: property "code" is equal to 18, expected 20
FAIL Given multiple nested browsing contexts, and window calls show() first, other nested browsing contexts can't show a request. promise_rejects_dom: Expected popupShowPromise to reject, request is already showing. function "function () { throw e }" threw object "SecurityError: show() must be triggered by user activation." that is not a DOMException AbortError: property "code" is equal to 18, expected 20
FAIL Given multiple nested browsing contexts, and popup calls show() first, other nested browsing contexts can't show a request. promise_rejects_dom: Expected windowShowPromise to reject, the popup is showing a payment request. function "function () { throw e }" threw object "SecurityError: show() must be triggered by user activation." that is not a DOMException AbortError: property "code" is equal to 18, expected 20
FAIL Given multiple nested browsing contexts, and an iframe calls show() first, other nested browsing contexts can't show a request. promise_rejects_dom: Expected windowShowPromise to reject, the popup is showing a payment request. function "function () { throw e }" threw object "SecurityError: show() must be triggered by user activation." that is not a DOMException AbortError: property "code" is equal to 18, expected 20
FAIL Navigating an iframe as a nested browsing context sets 'payment request is showing boolean' to false. promise_rejects_dom: Navigating iframe away must cause the iframeShowPromise to reject. function "function () { throw e }" threw object "SecurityError: show() must be triggered by user activation." that is not a DOMException AbortError: property "code" is equal to 18, expected 20
TIMEOUT Navigating a popup as a nested browsing context sets 'payment request is showing boolean' to false. Test timed out
